(feat) Downloads in temporary dir
This is specifically for Fargate, so downloads are _not_ in the same
shared volume as the watched directory. The shared volume has a limit of
4gb, but overall the limit is 10gb. This allows large files to be
redownloaded, say if there were changes.

It also offers fewer special cases in the code, since the download dir
does not overlap with the main directory