build(deps): remove dependency on lowhaio and move to httpx
The main reason for this is since lowhaio is suspected to hang. Since this is a
hefty change, to avoid more changes throughout the code, constructing a
lowhaio-like Pool object that actually uses httpx under the hood.

The change of hostname to netloc in signing is due to a suspected issue with
lowhaio - it does not include the port in the host header, when for non-default
ports, it should.

This also results in not using 0x20 encoding for resolving host names, so this
isn't technically just an implementation detail change. While potentially this
makes this a little bit less secure, it's not really that typical to be using
this, a bit beyond the scope of what this level of code "should" care about I
think, and it needs to be turned off when communicating with Docker containers
anyway.

DT-340