feat: reduce API calls in case where there is no folder object in files
If we have files in a folder (or prefix) but with no folder object in S3 (so
file ending in "/"), then there will be a needless API call to S3 for the
existance of a the root object. This will return a 404, and then an attempt
will be made to delete the local folder.

It's not the end of the world, but in some configs, too many 404s are
suspicious, so this should help reduce the noise.